chore: removed nowpage, homepage; simplified dir structure
